Build Windows Hybrid WireGuard v1.19.30 v2.2

+++ tools/hybrid/vendor_overlay/github.com/metacubex/gvisor/pkg/tcpip/network/ipv6/androidlike_flowlabel_test.go
@@ -3,7 +3,7 @@ package ipv6
 import "testing"
 
 func TestHybridIPv6DefaultHopLimit64(t *testing.T) {
-	if got, want := DefaultTTL, uint8(64); got != want {
+	if got, want := int(DefaultTTL), 64; got != want {
 		t.Fatalf("IPv6 default hop limit got=%d want=%d", got, want)
 	}
 }
